requirements.txt
-i https://pypi.org/simple
anyio==4.8.0; python_version >= '3.9'
blinker==1.9.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.6'
click==8.1.8; python_version >= '3.7'
flask==3.1.0; python_version >= '3.9'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.5; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
python-telegram-bot==21.10; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version < '3.13'
webvtt-py==0.5.1; python_version >= '3.7'
werkzeug==3.1.3; python_version >= '3.9'
yt-dlp==2025.2.19; python_version >= '3.9'
